Moved demo content plugins to match best practice.
