+++ src/Plugin/Presto/DemoContent/SetupCheckoutFlow.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\presto\Installer\Ecommerce\Content\Plugin;
+namespace Drupal\presto\Plugin\Presto\DemoContent;
 
 use Drupal;
 use Drupal\Core\Config\FileStorage;
@@ -14,7 +14,7 @@ use Drupal\Core\Config\FileStorage;
  *     weight = 4
  * )
  *
- * @package Drupal\presto\Installer\Ecommerce\Content\Plugin
+ * @package Drupal\presto\Plugin\Presto\DemoContent
  */
 class SetupCheckoutFlow extends AbstractDemoContent {
